add jedi, vimux, some syntax plugins
jedi might be messing with default ctrl+n auto completion.

default behavior jumps to the first match right away.

with jedi on it seems to only open the auto complete menu,
so you have to mash it twice every time.

must be an artifact of how jedi plugs into that dropdown.

need to figure out how to configure that so that
it's smart enough to be selective on which instances it hesitates