add lockscreen keybinding with hammerspoon
If you don't have the power button on your keyboard, you cannot lock
your screen.  Hot corners are garbage because you need to babysit the
cursor all the way, or sometimes your mouse doesn't work so it's
associating a gesture with a simple command.

By pegging it to a key you get an interface that works reliably.