make mojibar not interrupt layout and be in a little corner
this is one of those things that went from "wow the default behavior of
this application is completely unusable" to my finally figuring out how
to get i3 to do a simple thing to this program and feeling some weird
sense of accomplishment in spite of the fact that this just wouldn't
have been a problem if i were using a well-designed environment in the
first place.

that's the stockholm syndrome of linux, for ya.